moves topbase impl to seperate module
